app/fileupload_event_service_metadata/requirements.txt
## Requirements for trigger_function_file_metadata goes here.
#google-cloud-pubsub
google-cloud-firestore